refactor(aigateway): rename test support modules
Refs: OME-884
