+++ apps/aigateway/tests/unit/openai/ambient_state.py
