+++ apps/aigateway/tests/unit/openrouter/projection_harness.py
